Idea: Defining the shape of a Wrapper Scaffolding
The concept of "Wrapper Scaffolding" doens't exist perse but
here we are trying to define how should it looks like.

We are using the function 'scaffolding_load()' to load the parent
core go scaffolding first, all callbacks, variables and customizations.
Then it we will load this new wrapper scaffolding that will override
some of the functionality of the parent.

Added studio dev-env with helper methods to build and test this package.

To build a package run:

  # build

To build a simple Go application to test the latest scaffolding-go wrapper:

  # build_go_app

For faster development cycles use the following helper to build and test a
new scaffolding-go package:

  # build_and_test

Signed-off-by: Salim Afiune <afiune@chef.io>