Attempt to build and verify static binaries by default
This change makes `scaffolding_go_build()` attempt to build go binaries
that are statically linked. It does this by completely disabling CGO,
passing static flags to the external linker, and by applying go build
tags that enable the native go backends for the os/user and net
packages that might otherwise try to dynamically link to glibc.

Users of the scaffolding can disable this new functionality by setting
the scaffolding_go_no_static variable to a non-empty value in their
plan.

It also creates a default do_check hook that varifies that the any
binaries that are built have no dynamic links. The check can be disabled
using the aforementioned plan variable.

Signed-off-by: Ryan Cragun <ryan@chef.io>